+++ test/preload.py
@@ -7,13 +7,13 @@ sys.path = ['../../']+sys.path
 from expy import *  # Import the needed functions
 start()  # Initiate the experiment environment
 
-drawText('Hello', display=False)  # Draw text on the canvas
+drawText('Hello', show_now=False)  # Draw text on the canvas
 s1 = getScreen()  # Get current canvas, then clean the canvas
 
-drawText('world', display=False)  # Draw text on the canvas
+drawText('world', show_now=False)  # Draw text on the canvas
 s2 = getScreen(clean_screen=False)  # Get current canvas, and keep it
 
-drawText('........', display=False)  # Draw text on the canvas
+drawText('........', show_now=False)  # Draw text on the canvas
 s3 = getScreen()  # Get current canvas, then clean the canvas
 
 show(3, backup=s1)  # Display backup canvas