+++ test/show_picture.py
@@ -17,14 +17,14 @@ show(3)  # Display current canvas
 drawPic('data/demo.jpg', w=400, h=300)
 show(3)  # Display current canvas
 
-# # Draw a zoomed picture on the canvas center
-# drawPic('data/demo.jpg', w=300, h=400, rotate=90)
-# show(3)  # Display current canvas
+# Draw a zoomed picture on the canvas center
+drawPic('data/demo.jpg', w=300, h=400, rotate=90)
+show(3)  # Display current canvas
 
 # Draw a zoomed picture on the canvas, and move it
 drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5)
 show(3)  # Display current canvas
 
 # Draw a zoomed picture on the canvas, and move it
-drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5, anchor_x='left',anchor_y='center')
+drawPic('data/demo.jpg', w=400, h=300, x=0.5, y=0.5, anchor_x='right',anchor_y='center')
 show(3)  # Display current canvas