+++ test/time_delay.py
@@ -3,8 +3,26 @@
 import sys
 sys.path = ['../../']+sys.path
 ################
-
+import time
 from expy import *  # Import the needed functions
+
+# print('vsync=False')
+# start()
+# show(0.1)
+# for _ in range(10):
+#     now = time.time()
+#     shared.win.flip()
+#     print('%.5f' %(time.time() - now))
+# shared.win.close()
+
+# print('vsync=True')
+# start(vsync=True)
+# show(0.1)
+# for _ in range(10):
+#     now = time.time()
+#     shared.win.flip()
+#     print('%.5f' %(time.time() - now))
+# shared.win.close()
 
 # Zero-setting
 def reset():
@@ -13,44 +31,35 @@ def reset():
 
 def close():
     sendTrigger('mh70', mode='S') # close signal
-    shared.pa.terminate()
     shared.win.close()
     shared.ser.close()
     shared.pyglet.app.exit()
 
 def visualCases(count):
-    'draw run + flip - trigger send'
-    for _ in range(count):
-        reset()
-        sendTrigger('mh10', mode='S') # program trigger
-        drawText('Hello', display=False)
-        drawCircle(30, fill=True, x=-0.9, y=-0.9, display=False)
-        show(1) # physical trigger
-        show(0.5)
 
     'flip - trigger send'
     for _ in range(count):
         reset()
-        drawText('Hello', display=False)
-        drawCircle(30, fill=True, x=-0.9, y=-0.9, display=False)
-        sendTrigger('mh20', mode='S') # program trigger
+        drawText('Hello', show_now=False)
+        drawCircle(30, fill=True, x=-0.9, y=-0.9, show_now=False)
+        sendTrigger('mh10', mode='S') # program trigger
         show(1) # physical trigger
         show(0.5)
 
     'trigger run + trigger send - flip'
     for _ in range(count):
         reset()
-        drawText('Hello', display=False)
+        drawText('Hello', show_now=False)
         drawCircle(30, fill=True, x=-0.9, y=-0.9) # physical trigger
-        sendTrigger('mh30', mode='S') # program trigger
+        sendTrigger('mh20', mode='S') # program trigger
         show(1)
         show(0.5)
 
     'flip - trigger send'
     for _ in range(count):
         reset()
-        drawText('Hello', display=False)
-        drawCircle(30, fill=True, x=-0.9, y=-0.9,trigger=('mh40', 'S')) # program trigger, physical trigger
+        drawText('Hello', show_now=False)
+        drawCircle(30, fill=True, x=-0.9, y=-0.9,trigger=('mh30', 'S')) # program trigger, physical trigger
         show(1)
         show(0.5)
 
@@ -70,14 +79,18 @@ def auditoryCases(count):
         playSound(sound,trigger=('mh20', 'S'))
         show(0.5)
 
-start(port='COM5')
-visualCases(2)
-close()
-
-start(port='COM5',vsync=False)
-visualCases(2)
-close()
-
 start(port='COM5',sample_rate=44100)
+reset()
 auditoryCases(2)
 close()
+
+start(port='COM5')
+reset()
+visualCases(2)
+close()
+
+start(port='COM5',vsync=True)
+reset()
+visualCases(2)
+close()
+